requirements.txt
docker-py
requests
